+++ packages/expo-notifications/android/src/main/java/expo/modules/notifications/notifications/handling/SingleNotificationHandlerTask.java
@@ -5,6 +5,8 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.os.ResultReceiver;
 
+import androidx.core.os.BundleCompat;
+
 import expo.modules.kotlin.Promise;
 import expo.modules.core.interfaces.services.EventEmitter;
 import expo.modules.notifications.notifications.NotificationSerializer;
@@ -98,7 +100,7 @@ public class SingleNotificationHandlerTask {
             if (resultCode == NotificationsService.SUCCESS_CODE) {
               promise.resolve();
             } else {
-              Exception e = (Exception) resultData.getSerializable(NotificationsService.EXCEPTION_KEY);
+              Exception e = BundleCompat.getSerializable(resultData, NotificationsService.EXCEPTION_KEY, Exception.class);
               promise.reject("ERR_NOTIFICATION_PRESENTATION_FAILED", "Notification presentation failed.", e);
             }
           }